[add] builtins function with a function pointer tab

+++ src/str/line_editing.c
@@ -78,7 +78,7 @@ int print_line(char *command, int *index, char const *prompt_char)
     printf("\r");
     if (index[0] == -2)
         index[0]++;
-    for (int i = 0; i < index[0] +  strlen(prompt_char) + 1; i++)
+    for (int i = 0; i < index[0] +  (int)strlen(prompt_char) + 1; i++)
         printf("\033[C");
     return 0;
 }
